[artf1202305][AOSP]:Check flags when receiving INTF_ACTIVATED_NTF in RFST_W4_HOST_SELECT
If the flag W4_RSP is set when receiving an INTF_ACTIVATED_NTF while
in RFST_W4_HOST_SELECT, this means an RF_DEACTIVATE_CMD was sent just
before the NTF is received and has not yet been processed.
In this case the RFST changes to RFST_POLL/LISTEN_ACTIVE and in that
state, the flag W4_NTF must also be set.

Bug: 340223581
Test: manual

URL:https://android-review.googlesource.com/c/platform/system/nfc/+/3128014?forceReload=true