[artf1202305][AOSP]: In RFST_IDLE, block RF_DISCOVER_RSP callback only if DISABLING
If an RF_DEACTIVATE_CMD is sent just after an RF_DISCOVER_CMD
when in RFST_IDLE, the command should be sent and a callback done
to the JNI to unblock any waiting process/mutex. The only case when
this callback is not needed is when the stack is disabling.

Bug: 374687433
Test: manual

URL: https://android-review.googlesource.com/c/platform/system/nfc/+/3128026?forceReload=true