[artf1202305][AOSP]: nfc: Fix a bug in RegisteredServicesCache
The bug is that the uid is not correctly set when creating a new DynamicSettings object from within the NFC stack (migration).
This can cause problems when the object is later used.

Also, removed uid checking on parsing which is redundant and will
help users upgrade to fix the issue.

Bug: 379683345
Test: Manual verification

URL: https://android-review.googlesource.com/c/platform/packages/apps/Nfc/+/3362967?forceReload=true