[artf1209570][AOSP_AUTO]: Added NFA_SetNfcSecure() API
This is needed when setNfcSecure() is called:

only power state screen_on_unlocked is allowed in this case, but in

addition all other power states should be blocked. If not, the NFCC

will have no matching entries for the non supported power states, and

it will be up to NFCC to decide how to route. It might route the CE

to an host that will answer.

Bug: 378474016

Test: manualURL: https://android-review.googlesource.com/c/platform/system/nfc/+/3262076?forceReload=true