[artf1209570][AOSP_AUTO]: Do not allocate/deallocate twice AID buffers to avoid crash
When called, NFA_EeRegister() will register a callback for the calling

entity and perform the allocation of the buffers for the AID management.

If called twice, then this operation is done again.

The code allocating the buffers has been moved to nfa_ee_sys_enable().

When nfa_ee_sys_disable() is called at NFC disable, nfa_ee_api_deregister()

is called for as many callbacks that were registered.

The function will reset the registered callback function and

free the buffers allocated for AID management.

If called twice, the 2nd call will try to free buffers already freed,

leading to a crash. This code has been moved to nfa_ee_sys_disable().

Bug: 378533480

Test: manualURL: https://android-review.googlesource.com/c/platform/system/nfc/+/3267499?forceReload=true