[artf1209570][AOSP_AUTO]: Add flag in nfa_ee_cb.ee_cfged for default AID route
Default AID route now has its own entry in nfa_ee_cb.ecb at index 5.

When default AID route is set, only bit 5 of nfa_ee_cb.ee_cfged is set.

The bit corresponding to the route for the AID is not set.

If no other type of routing is done on the route for AID, then when

calling nfa_ee_need_recfg(), it will not detect the RT needs to be

updated.

Bug: 378533480

Test: manualURL: https://android-review.googlesource.com/c/platform/system/nfc/+/3267500?forceReload=true