[artf1209570][AOSP_AUTO]: RoutingManager.cpp rework: using DH as fallback route
According to NCI, if the route resolution procedure ends with "unknown:

"If no route is found for the received frame, the NFCC handles

the frame in an implementation-specific manner."

As this depends on NFCC implementation, the outcome might lead to

unwanted and nonlinear behaviors.

Routing everything to DH if the requested route is not available/relevant

ensure the result of the routing process is never "unknown".

Bug: 378643826

Test: Manual

URL: https://android-review.googlesource.com/c/platform/packages/modules/Nfc/+/3439133?forceReload=true