[artf1265062][25Q4_AUTO]: [owners] Remove alisher@google.com from OWNERS
This suggested change is automatically generated based on group

memberships and affiliations.

Please approve this change and vote the highest CR. This will keep the

OWNERS file tidy. We ask that you do not ignore this change and either

approve it if the OWNER is no longer active or deny it if the OWNER is

continuing to review CLs (including under a different affiliation).

This CL can always be reverted if the understanding changes.

If this change is in error, vote the lowest CR value (i.e. reject the

CL) and the bot will abandon it.

See the owner's recent review activity for context:

https://googleplex-android-review.googlesource.com/q/alisher@google.com

To report an issue, file a bug in the Infra>Codereview component.