Fix ubsan issue with older icode tags
For older ICODE tags (no support GetMultiBlockSecurityStatus),

the stack checks the security status of the blocks one by one to determine

the maximum writable area for NDEF.

However if the first read-only block is within the NDEF TLV area,

the operation used causes an overflow that was triggering an exception

with the sanitize compilation options.

We add a check on the offset of this first RO block to ensure the operation

is correct.

Bug: 419601692

Test: tag reading

Flag: EXEMPT bug bix