[artf1307670][26Q2_AUTO]: Replace {@hide} with @hide
In most places the `@hide` doctag is treated as a block tag on its own

line. However, there are a few places where it is treated as an inline

doctag, i.e. `{@hide}`. Dealing with both types increases the

complexity of Metalava and makes it slightly slower.

This change replaces the inline doctag form `{@hide}` with a block tag.

Flag: EXEMPT PURE_REFACTOR

      replacing {@hide} with @hide does not change API or behavior

Bug: 429965593

Test: m checkapi