[artf1307670][26Q2_AUTO]: NFC: Add Null-dereference READ in rw_t1t_process_error
The key changes are wrapping the retry logic in an if (p_t1t->p_cur_cmd_buf != nullptr) block and adding an else to log an error if the pointer is null. This prevents a potential crash from dereferencing a null pointer in the memcpy function.

Bug: 447923856

Test: N/A

Flag: EXEMPT BUGFIX