[artf1311199][26Q2_AUTO]: Fix the three failing tests in `NfcNciUnitTests`.
Originally all failing tests were in HostNfcFEmulationManagerTest. The

cause is that `Message.obtain()` should not be mocked in a way that

returns a fixed `message` all the time: it's likely that one part of

the code sets a callback on the message for its own internal logic. If

the same object is always returned, the callback might be unexpected.

Bug: 455155203

Flag: TEST_ONLY

Test: atest NfcNciUnitTests

Change-Id: I814d72d76d3020a70c50607700a9c6e15058fe58